+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Add `EveOnline::ESI::CharacterImplants`
 * Drop `EveOnline::XML::CharacterAccountBalance`. Use `EveOnline::ESI::CharacterWallet` instead.
 * Drop `EveOnline::XML::CharacterSheet#balance`. Use `EveOnline::ESI::CharacterWallet` instead.
+* Drop `EveOnline::XML::CharacterSheet#implants`. Use `EveOnline::ESI::CharacterImplants` instead.
 
 **v0.12.0**